Add: Logica de compra de servicios en las ciudades
- Ya se puede comprar un servicio en la ciudad donde esta la carvana.
- La compra del servicio resta dinero.
- La compra del servicio verifica que la caravana no la haya hecho previamente y que cumpla los requisitos a nivel de stats.

+++ Juego-Caravana-Medieval/src/main/java/co/edu/javeriana/caravana_medieval/dto/CaravanaDTO.java
@@ -7,14 +7,14 @@ public class CaravanaDTO {
     private String nombre;
     private int velocidad;
     private int capacidadMax;
-    private int dineroDisponible;
+    private double dineroDisponible;
     private int puntosVida; 
     private LocalTime horaViaje; 
 
     public CaravanaDTO() {
     }
 
-    public CaravanaDTO(Long id,String nombre, int velocidad, int capacidadMax, int dineroDisponible, int puntosVida, LocalTime horaViaje) {
+    public CaravanaDTO(Long id,String nombre, int velocidad, int capacidadMax, double dineroDisponible, int puntosVida, LocalTime horaViaje) {
         this.id = id;
         this.nombre = nombre;
         this.velocidad = velocidad;
@@ -64,11 +64,11 @@ public class CaravanaDTO {
         this.capacidadMax = capacidadMax;
     }
 
-    public int getDineroDisponible() {
+    public double getDineroDisponible() {
         return dineroDisponible;
     }
 
-    public void setDineroDisponible(int dineroDisponible) {
+    public void setDineroDisponible(double dineroDisponible) {
         this.dineroDisponible = dineroDisponible;
     }
